Refactored tests structure
