+++ cypress/support/selectors.ts
@@ -1,0 +1,3 @@
+export const schemaEditorLabel = "current schema editor";
+export const generatedModelLabel = "generated model";
+export const suggestionSelector = `[widgetid="editor.widget.suggestWidget"][monaco-visible-content-widget="true"]`;
